First Commit
Establishes app.js, index.html and main.css. Currently there is minimal
functionality in the app, just very simple UI.

Tag Input Displays, Tag Chips are created when a tag is added. The
tags[] array is created and is observable. Remove tag works by removing
the last inputed tag, the chip clears and it is removed from the array.