Long Term Fix
Reverted gitnore to include the bower-module directory and node-module
directory.
Updated the readme to include instructions on installing node and bower
dependencies.

This should reduce the size of the repository considerably and allow for
faster more reliable syncing across git. readme still needs some
instructions to be updated.

In the long term Gulp, Bower and NPM package files should be built out
to simplify dependency installation.